# doc/api.yaml
openapi: 3.0.0
info:
  title: wasa-Photo
  description: |
    Each user will be presented with a stream of photos in reverse chronological order, with information
    about when it was uploaded (date and time) and how many likes and comments it has. Users can place
    (and later remove) a “like” to photos from other users. Also, users can add comments to any photo
    (even those uploaded by themself). Comments can be removed by the author only.
    Users can ban other users. If Alice (user) bans Eve (user), Eve won’t be able to see any information
    about Alice. Alice can decide to remove the ban at any moment.
    Users will have their own personal profiles. The personal profile page for user detail: user’s photos (in
    reverse chronological order), their followers and following, and how many photos have been uploaded.
    Each user can change his/her own username, upload photos, remove photos, and follow/unfollow
    other users. Removal of a photo will also remove likes and comments.
    User profiles can be searched via username.
  version: 0.0.1
tags:
  - name: User
    description: Operations about user
  - name: Photos
    description: Operations about photo
  - name: Other User
    description: Operations about other user
  - name: Login
    description: Operations about login




paths:  
  /session:
    post:
      tags: ["Login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Maria"
                  pattern: '^[a-zA-Z0-9_]*$'
                  minLength: 3
                  maxLength: 64
        required: true
      responses:
        '201':
          description: |
            User log-in action successful returning session ID
          content:  
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    minLength: 1
                    maxLength: 128
                    pattern: '^[a-z0-9]*$'
                    example: "2d38da438c252b0d5755e1c1e8caa104"
  /home:
        get:
          tags: ["User"]
          summary: Get user home page
          description: |-
            Get the  user home page
          operationId: getMyStream
          responses:
            "200":
              description: "user Home page loaded correcttly"
              content:
                application/json:
                  schema:
                    type: array
                    items:
                      type: object
                      properties:
                        photo:
                          $ref: '#/components/schemas/Photo'
                    maxItems: 1024
                    minItems: 0
  /users:
    get:
       tags: ["User"]
       summary: "get users , either all or searching for a username"
       description: |
            Get the users list if no parameter given, otherwise search for some users
       operationId: getUsers
       parameters:
        - name : otherusername
          in: query
          description: "other user name"
          required: true
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]*$'
            minLength: 3
            maxLength: 64
            example: lucas
       responses:
        "200":
          description: "users list found "
          content:  
              application/json:
                schema:
                  type: array
                  items: 
                    type: string
                    pattern: "^[a-zA-Z0-9_]*$"
                    minLength: 3
                    maxLength: 64
                  minItems: 0  
                  maxItems: 1024
                  example: 
                    -"Maria"
                    -"John"
  /users/{username}:
    get:
      tags: ["User"]
      summary: "Get user information"
      description: |-
        returns back a j-son file object containing all the information
        related to that user Id
      operationId: getUserProfile
      parameters:
        - $ref: '#/components/parameters/username'

      responses:
        "200":
          description: 
            "user correctly retrieved "
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{username}/options:
    put:
      parameters:
        - $ref: '#/components/parameters/username'
      tags: ["User"]
      summary: "Change username"
      description: |-
        Change the user name
      operationId: setMyUserName
      requestBody:
        description: User New Username
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string 
                  pattern: "^[a-zA-Z0-9_]*$"
                  minLength: 3
                  maxLength: 64
              example: "lucas2001"
        required: true
      responses:
        "204":
          description: 
            "user name changed"
  /users/{username}/following/{otherusername}:
    parameters:
        - $ref: "#/components/parameters/othersusername"
        - $ref: "#/components/parameters/username"
    post:
      tags: [Other User]
      summary: "starts following a user "
      description: |-
        it modifies the following list of the logged in user
        adding the selected user
      operationId: followUser
      responses:
        "201":
          description: "user correctly added to the following list"
    delete:
      tags: [Other User]
      summary: "unfollows a user"
      description: | 
        "it modifies the following list of the logged in user 
        removing the selected user "   
      operationId: unfollowUser
      responses:
        "204":
          description: "user correctly removed from the following list"
  /users/{username}/banned/{otherusername}: 
    parameters:
        - $ref: "#/components/parameters/othersusername"
        - $ref: "#/components/parameters/username"
    post: 
      tags: [Other User]
      summary: "bans a user"
      description: "takes the id of a user and adds it to the banned list of another user "
      operationId: "banUser"
      responses:
        "201":
          description: "user correctly added to the banned list"
    delete:
      tags: [Other User]
      summary: "unbans a user"
      description: "takes the id of a user and removes it from the banned list of another user"
      operationId: "unbanUser"
      responses:
        "204":
          description: "user correctly removed from the banned list"
  /users/{username}/photos:
    parameters:
      - $ref: "#/components/parameters/username"
    post:
      tags : ["Photos"]
      summary: "post a new photo"
      description: "gets a photo from the client and create a resource photo{id} in the server"
      operationId: "uploadPhoto"
      parameters:
        - name: "photo"
          in : query
          required: true
          schema :
            type: string
            pattern: 'data:image\/[^;]+;base64[^"]+'
            minLength: 32
            maxLength: 1000000
            example: |
              "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABgAAAAYCAYAAADgdz34AAAABHNCSVQICAgIfAhkiAAAAAlwSFlzAAAApgAAAKYB3X3/
              OAAAABl0RVh0U29mdHdhcmUAd3d3Lmlua3NjYXBlLm9yZ5vuPBoAAANCSURBVEiJtZZPbBtFFMZ/M7ubXdtdb1xSFyeilBapySVU8h8OoFaooFSqiihIVIpQBKci6KEg9Q6H9kovIHoCIVQJJCKE1ENFjnAgcaSGC6rEnxBwA04Tx43t2FnvDAfjkNibxgHxnWb2e
              /u992bee7tCa00YFsffekFY+nUzFtjW0LrvjRXrCDIAaPLlW0nHL0SsZtVoaF98mLrx3pdhOqLtYPHChahZcYYO7KvPFxvRl5XPp1sN3adWiD1ZAqD6XYK1b/dvE5IWryTt2udLFedwc1+9kLp+vbbpoDh+6TklxBeAi9TL0taeWpdmZzQDry0AcO+jQ12RyohqqoYoo8RDwJrU+qXkjWtfi8Xxt58BdQuwQs9qC
              /afLwCw8tnQbqYAPsgxE1S6F3EAIXux2oQFKm0ihMsOF71dHYx+f3NND68ghCu1YIoePPQN1pGRABkJ6Bus96CutRZMydTl+TvuiRW1m3n0eDl0vRPcEysqdXn+jsQPsrHMquGeXEaY4Yk4wxWcY5V
              /9scqOMOVUFthatyTy8QyqwZ+kDURKoMWxNKr2EeqVKcTNOajqKoBgOE28U4tdQl5p5bwCw7BWquaZSzAPlwjlithJtp3pTImSqQRrb2Z8PHGigD4RZuNX6JYj6wj7O4TFLbCO/Mn/m8R+h6rYSUb3ekokRY6f/YukArN979jcW+V
              /S8g0eT/N3VN3kTqWbQ428m9/8k0P/1aIhF36PccEl6EhOcAUCrXKZXXWS3XKd2vc/TRBG9O5ELC17MmWubD2nKhUKZa26Ba2+D3P+4/MNCFwg59oWVeYhkzgN/JDR8deKBoD7Y+ljEjGZ0sosXVTvbc6RHirr2reNy1OXd6pJsQ+gqjk8VWFYmHrwBzW/n+uMPFiRwHB2I7ih8ciHFxIkd/3Omk5tCDV1t+2nNu5sxxpDFNx+huNhVT3
              /zMDz8usXC3ddaHBj1GHj/As08fwTS7Kt1HBTmyN29vdwAw+/wbwLVOJ3uAD1wi/dUH7Qei66PfyuRj4Ik9is+hglfbkbfR3cnZm7chlUWLdwmprtCohX4HUtlOcQjLYCu+fzGJH2QRKvP3UNz8bWk1qMxjGTOMThZ3kvgLI5AzFfo379UAAAAASUVORK5CYII="
      responses:
        "201":
          description: "picture uploaded correctly, returning the id of that picture"
          content:
            application/json:
              schema:
                type: string
                pattern: "^[a-z0-9_]*$"
                minLength: 1
                maxLength: 1024
                example: ejasorjtd8574
  /users/{username}/Photos/{photoid}:
    parameters:
        - $ref: "#/components/parameters/pothosid"
        - $ref: "#/components/parameters/username"
    get:
      tags: [Photos]
      summary: "Get a photo"
      description: |-
        returns the photo object with the given id
      operationId: getPhoto
      responses:
        "202":
          description: "photo correctly returned"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        "404":
          description: "photo not found"
    delete:
      tags: [Photos]
      summary: "delete a picture"
      description: |
        delete a picture from the server
      operationId: "deletePhoto"
      responses:
        "204":
          description: "picture deleted correctly"
  /users/{username}/Photos/{photoid}/likes:
    parameters:
    - $ref: "#/components/parameters/pothosid"
    - $ref: "#/components/parameters/username"
    delete:
        tags: [Photos]
        summary: "delete a like"
        description: |
          delete a like from a picture
        operationId: "unlikePhoto"
        responses:
          "204":
            description: "like deleted correctly"
    put:
      tags: ["Photos"]
      summary: "like a picture"
      description: "takes the Id of a Picture , the ID of the logged in user , and adds a like to it"
      operationId: "likePhoto"
      responses:
        "201":
          description: "liked added correctly"
          content:
            application/json:
              schema:
                description:  "the like id of that like "
                type: integer
                example: 12
  /users/{username}/Photos/{photoid}/comments:
      parameters:
      - $ref: "#/components/parameters/pothosid"
      - $ref: "#/components/parameters/username"
      post: 
        tags: [Photos]   
        summary: "add a comment to a picture"
        description: |
          "takes the Id of a Picture ,
           the token of the logged in user , 
           and adds a comment to it"   
        operationId: "commentPhoto"
        parameters:
          - name: "comment"
            in : query
            required: true
            schema :
              type: string
              minLength: 1
              maxLength: 2048
              pattern: '[A-Za-z0-9_@./#&+-?=$%&()_^]$'
              example: "Wow wonderfull picture"
        responses:
          "201":
            description: "comment added correctly"
            content:
              application/json:
                schema:
                  type: integer
                  example: 10
  /users/{username}/Photos/{photoid}/comments/{commentid}:
     parameters:
       - $ref: "#/components/parameters/pothosid"
       - $ref: "#/components/parameters/username"
       - $ref: "#/components/parameters/commentid"
     delete:
          tags: [Photos]
          summary: "delete a comment"
          operationId: uncommentPhoto
          description: |
            "takes the Id of a Picture ,
            the token of the logged in user , 
            the id of the comment to delete,
            and deletes a comment from it"
          responses:
            "204":
              description: "comment deleted correctly"
components: 
  parameters:
    username:
      name: username
      in: path
      description: "user name"
      required: true
      schema:
        type: string
        minLength: 3
        maxLength: 64
        pattern: "^[a-zA-Z0-9_]*$"
        example: lucas
    othersusername:
      name: otherusername
      in: path
      description: "other user name"
      required: true
      schema:
        type: string
        minLength: 3
        maxLength: 64
        pattern: "^[a-zA-Z0-9_]*$"
        example: lucas
    pothosid:
      name: photoid
      in: path 
      description: "photo id"
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 1024
        pattern: "^[a-z0-9]*$"
        example: ejasorjtd8574
    commentid:
      name: commentid
      in: path
      description: "comment id"
      required: true
      schema:
        type: integer
        example: 10
  schemas:
    User:
      description:  "user object"
      type: object
      properties:
          followers:
            description: "followers of a user"
            type: array
            maxItems: 4294967296
            minItems: 0
            items:
              type: string
              minLength: 3
              maxLength: 64
              pattern: "^[a-zA-Z0-9_]*$"
              example: lucas
          following:
            description: "following  of a user"
            type: array
            maxItems: 4294967296
            minItems: 0
            items:
                type: string
                minLength: 3
                maxLength: 64
                pattern: "^[a-zA-Z0-9_]*$"
                example: lucas
            example:
              - jhonny 
              - jacob
              - cindy 
          number-of-photos :
            description: "number of photos of a user"
            type: integer
            example: 4
          photos: 
            description: "photos of a user"
            type: array
            maxItems: 4294967296
            minItems: 0
            items:
                type: string
                minLength: 1
                maxLength: 1024
                pattern: "^[a-z0-9]*$"
                example: ejasorjtd8574
    Photo:
      description: "photo object"
      type: object
      properties:
        image:
          description: "image of a photo"
          type: string
          pattern: 'data:image\/[^;]+;base64[^"]+'
          maxLength: 1000000
          minLength: 32
          example: |
              "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABgAAAAYCAYAAADgdz34AAAABHNCSVQICAgIfAhkiAAAAAlwSFlzAAAApgAAAKYB3X3/
              OAAAABl0RVh0U29mdHdhcmUAd3d3Lmlua3NjYXBlLm9yZ5vuPBoAAANCSURBVEiJtZZPbBtFFMZ/M7ubXdtdb1xSFyeilBapySVU8h8OoFaooFSqiihIVIpQBKci6KEg9Q6H9kovIHoCIVQJJCKE1ENFjnAgcaSGC6rEnxBwA04Tx43t2FnvDAfjkNibxgHxnWb2e
              /u992bee7tCa00YFsffekFY+nUzFtjW0LrvjRXrCDIAaPLlW0nHL0SsZtVoaF98mLrx3pdhOqLtYPHChahZcYYO7KvPFxvRl5XPp1sN3adWiD1ZAqD6XYK1b/dvE5IWryTt2udLFedwc1+9kLp+vbbpoDh+6TklxBeAi9TL0taeWpdmZzQDry0AcO+jQ12RyohqqoYoo8RDwJrU+qXkjWtfi8Xxt58BdQuwQs9qC
              /afLwCw8tnQbqYAPsgxE1S6F3EAIXux2oQFKm0ihMsOF71dHYx+f3NND68ghCu1YIoePPQN1pGRABkJ6Bus96CutRZMydTl+TvuiRW1m3n0eDl0vRPcEysqdXn+jsQPsrHMquGeXEaY4Yk4wxWcY5V
              /9scqOMOVUFthatyTy8QyqwZ+kDURKoMWxNKr2EeqVKcTNOajqKoBgOE28U4tdQl5p5bwCw7BWquaZSzAPlwjlithJtp3pTImSqQRrb2Z8PHGigD4RZuNX6JYj6wj7O4TFLbCO/Mn/m8R+h6rYSUb3ekokRY6f/YukArN979jcW+V
              /S8g0eT/N3VN3kTqWbQ428m9/8k0P/1aIhF36PccEl6EhOcAUCrXKZXXWS3XKd2vc/TRBG9O5ELC17MmWubD2nKhUKZa26Ba2+D3P+4/MNCFwg59oWVeYhkzgN/JDR8deKBoD7Y+ljEjGZ0sosXVTvbc6RHirr2reNy1OXd6pJsQ+gqjk8VWFYmHrwBzW/n+uMPFiRwHB2I7ih8ciHFxIkd/3Omk5tCDV1t+2nNu5sxxpDFNx+huNhVT3
              /zMDz8usXC3ddaHBj1GHj/As08fwTS7Kt1HBTmyN29vdwAw+/wbwLVOJ3uAD1wi/dUH7Qei66PfyuRj4Ik9is+hglfbkbfR3cnZm7chlUWLdwmprtCohX4HUtlOcQjLYCu+fzGJH2QRKvP3UNz8bWk1qMxjGTOMThZ3kvgLI5AzFfo379UAAAAASUVORK5CYII="
 

          
        id:
          description: "id of a photo"
          type: string
          minLength: 1
          maxLength: 1024
          pattern: "^[a-z0-9]*$"
          example: ejasorjtd8574
        likes:
          description: "likes of a photo"
          type: array
          maxItems: 4294967296
          minItems: 0
          items:
            type: string
            minLength: 3
            maxLength: 64
            pattern: "^[a-zA-Z0-9_]*$"
            example: lucas
        comments:
          description: "comments of a photo"
          type: array
          maxItems: 4294967296
          minItems: 0
          items:
            type: string
            minLength: 1
            maxLength: 2048
            pattern: '^[A-Za-z0-9_@./#&+-?=$%&()_ ^]*$'
            example: "Wow wonderfull picture"
        date:
          description: "date of the photo"
          type: string
          format: date
          example: "2020-12-12"
  securitySchemes:
    bearerAuth:    
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []